rupees migration: peripheral columns + backend /100 sweep + Decimal JSON
Backend:
- Rewards vouchers + printer cost matrix -> rupees (already committed
  separately as 8d2ee28; this commit covers everything after)
- app/core/money.py + database/session.py: SQLAlchemy engine now uses a
  Decimal-safe JSON serializer so JSON columns (e.g. audit_logs before/
  after state) don't crash when a snapshot contains a money field
- auditlog/service.py: recursively sanitize Decimal -> float before
  writing before_state/after_state/metadata (works regardless of engine)
- Fixed real int() truncation bugs (silent precision loss on real rupee
  values) in group_cart/service.py, orders/order_service.py,
  slots/combined_service.py, admin ledger-adjustment endpoint
- Fixed a critical, previously-unknown class of bug: FastAPI routes with
  an explicit return type annotation (e.g. -> dict[str, Any]) serialize
  Decimal via Pydantic, which defaults to a JSON STRING ("120.00") not a
  number -- bypassing the app-level ENCODERS_BY_TYPE override. Fixed ~40
  call sites across admin/cart/fraud/group_cart/menu/orders/payments/
  recommendations/stationery/vendors by explicitly casting to float()
  before inserting a money value into a response dict
- Widened menu/stationery Form + Pydantic schema types int -> float so
  decimal rupee prices aren't rejected/truncated at the API boundary
- price_paise -> price rename in AI/recommendation schemas + smart_engine

Frontend (user/vendor/admin):
- Each app's formatPaise/formatMoneyPaise helper flipped from dividing
  by 100 to a pass-through (API now returns rupees everywhere); fixed
  the resulting double-conversion sites that had pre-multiplied by 100
  to compensate for the old behavior
- Renamed stale _paise-suffixed API fields end-to-end where still used
  (vouchers, printer pricing, admin analytics/trends)
- Fixed reverse-direction bugs: forms that multiplied user rupee input
  by 100 before POSTing (menu/stationery price editors, admin print-cost
  and ledger-adjustment forms)

Verification:
- All 3 frontends compile with zero TypeScript errors
- Backend boots clean; full pytest suite (928 collected, excluding one
  file that crashes on an unrelated pre-existing python-magic/Windows
  DLL issue) run against both this branch and the pre-migration master
  baseline: 128 failed / 943 passed / 16 skipped on BOTH, with an
  IDENTICAL failure list -- confirms zero regressions from this change

+++ tnt-backend-main/app/modules/menu/router.py
@@ -34,7 +34,7 @@ router = APIRouter(prefix="/menu", tags=["Menu & Inventory"])
 @router.post("/items", response_model=MenuItemResponse)
 def add_menu_item(
     name: str = Form(...),
-    price: int = Form(...),
+    price: float = Form(...),
     description: str | None = Form(None),
     category: str = Form("food"),
     prep_time_minutes: int | None = Form(None),
@@ -101,7 +101,7 @@ def get_menu_item_by_id(item_id: int, db: Session = Depends(get_db)):
 def edit_menu_item(
     item_id: int,
     name: str | None = Form(None),
-    price: int | None = Form(None),
+    price: float | None = Form(None),
     description: str | None = Form(None),
     is_available: bool | None = Form(None),
     prep_time_minutes: int | None = Form(None),
@@ -305,7 +305,7 @@ def add_stationery_service(
     service_type: str = Form(...),
     name: str = Form(...),
     description: str | None = Form(None),
-    price_per_page: int = Form(...),
+    price_per_page: float = Form(...),
     max_capacity: int | None = Form(None),
     is_available: bool = Form(True),
     db: Session = Depends(get_db),
@@ -359,7 +359,7 @@ def edit_stationery_service(
     service_id: int,
     name: str | None = Form(None),
     description: str | None = Form(None),
-    price_per_page: int | None = Form(None),
+    price_per_page: float | None = Form(None),
     max_capacity: int | None = Form(None),
     current_load: int | None = Form(None),
     is_available: bool | None = Form(None),
